Consolidate Db*Ops traits into single ops module
Move DbReadOps, DbWriteOps, DbTransactionOps, DbMetadataOps, and
DbCacheManagerOps into slatedb/src/ops.rs. Also expose WriteHandle::new
publicly. Public API at slatedb::* is unchanged.
