+++ slatedb/src/lib.rs
@@ -45,15 +45,11 @@ pub use compactor_state::VersionedCompactions;
 pub use config::{Settings, SstBlockSize};
 pub use db::{Db, DbBuilder, DbReaderBuilder, DbStatus, WriteHandle};
 pub use db_cache::stats as db_cache_stats;
-pub use db_cache_manager::{CacheTarget, DbCacheManagerOps};
+pub use db_cache_manager::CacheTarget;
 pub use db_iter::DbIterator;
-pub use db_metadata::DbMetadataOps;
-pub use db_read::DbReadOps;
 pub use db_reader::DbReader;
 pub use db_snapshot::DbSnapshot;
 pub use db_transaction::DbTransaction;
-pub use db_transaction_ops::DbTransactionOps;
-pub use db_write::DbWriteOps;
 pub use error::{CloseReason, Error, ErrorKind};
 pub use filter::BloomFilter;
 pub use filter_policy::{BloomFilterPolicy, Filter, FilterBuilder, FilterPolicy, FilterQuery};
@@ -63,6 +59,7 @@ pub use garbage_collector::GarbageCollectorBuilder;
 pub use iter::IterationOrder;
 pub use manifest::VersionedManifest;
 pub use merge_operator::{MergeOperator, MergeOperatorError};
+pub use ops::{DbCacheManagerOps, DbMetadataOps, DbReadOps, DbTransactionOps, DbWriteOps};
 pub use prefix_extractor::{PrefixExtractor, PrefixTarget};
 pub use rand::DbRand;
 #[cfg(test)]
@@ -113,15 +110,11 @@ mod db;
 mod db_cache_manager;
 mod db_common;
 mod db_iter;
-mod db_metadata;
-mod db_read;
 mod db_reader;
 mod db_snapshot;
 mod db_state;
 mod db_status;
 mod db_transaction;
-mod db_transaction_ops;
-mod db_write;
 mod dispatcher;
 mod error;
 pub mod filter;
@@ -138,6 +131,7 @@ mod memtable_flusher;
 mod merge_iterator;
 mod merge_operator;
 mod object_stores;
+mod ops;
 mod oracle;
 mod partitioned_keyspace;
 mod paths;
